# modulo5/Cliente-Swagger/api-docs.json
{"swagger":"2.0","info":{"description":"Api Clients","version":"1.0","title":"Api Clients","termsOfService":"Terminos del servicio","contact":{"name":"Pablo","url":" "},"license":{"name":"Apache License Version 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.txt"}},"host":"localhost:8080","basePath":"/clientes/v1","tags":[{"name":"cliente-controller","description":"Cliente Controller"}],"paths":{"/clientes":{"get":{"tags":["cliente-controller"],"summary":"return all clients","operationId":"findAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Cliente"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["cliente-controller"],"summary":"create a client","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"cliente","description":"cliente","required":true,"schema":{"$ref":"#/definitions/Cliente"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/clientes/apellido/{apellido}":{"get":{"tags":["cliente-controller"],"summary":"return a single client by surname","operationId":"getBySurnameUsingGET","produces":["*/*"],"parameters":[{"name":"apellido","in":"path","description":"apellido","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/clientes/correo/{correo}":{"get":{"tags":["cliente-controller"],"summary":"return a single client by email","operationId":"getByEmailUsingGET","produces":["*/*"],"parameters":[{"name":"correo","in":"path","description":"correo","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/clientes/{id}":{"get":{"tags":["cliente-controller"],"summary":"return a single client by id","operationId":"getByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["cliente-controller"],"summary":"update a client by id","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"cliente","description":"cliente","required":true,"schema":{"$ref":"#/definitions/Cliente"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["cliente-controller"],"summary":"delete a client","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Cliente"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"Cliente":{"type":"object","properties":{"_links":{"type":"array","items":{"$ref":"#/definitions/Link"}},"apellido":{"type":"string"},"correo":{"type":"string"},"direccion":{"type":"array","items":{"type":"string"}},"dni":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"telefono":{"type":"string"}},"title":"Cliente"},"Link":{"type":"object","properties":{"deprecation":{"type":"string","xml":{"name":"deprecation","attribute":true,"wrapped":false}},"href":{"type":"string","xml":{"name":"href","attribute":true,"wrapped":false}},"hreflang":{"type":"string","xml":{"name":"hreflang","attribute":true,"wrapped":false}},"media":{"type":"string","xml":{"name":"media","attribute":true,"wrapped":false}},"rel":{"type":"string","xml":{"name":"rel","attribute":true,"wrapped":false}},"templated":{"type":"boolean"},"title":{"type":"string","xml":{"name":"title","attribute":true,"wrapped":false}},"type":{"type":"string","xml":{"name":"type","attribute":true,"wrapped":false}}},"title":"Link"}}}